Split runtime module by concern
